cleanup of the source code, improve portability
