+++ run.cmd
@@ -1,0 +1,3 @@
+echo How to start marking bounging boxes to create training sets for Yolo
+
+yolo_mark.exe data/img data/train.txt data/obj.names
