+++ train.cmd
@@ -1,0 +1,3 @@
+echo How to start training Darknet Yolo objects (air, bird)
+
+darknet.exe detector train data/obj.data yolo-obj.cfg darknet19_448.conv.23
