--- x64/Release/train_obj.cmd
@@ -1,7 +1,0 @@
-echo     Example how to start training Darknet Yolo v2 for 2 objects (air, bird)
-
-
-darknet.exe detector train data/obj.data yolo-obj.cfg darknet19_448.conv.23
-
-
-pause
