--- x64/Release/yolo_mark.cmd
@@ -1,7 +1,0 @@
-echo     Example how to start marking bouded boxes for training set Yolo v2
-
-
-yolo_mark.exe data/img data/train.txt data/obj.names
-
-
-pause
