--- yolo_mark.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "yolo_mark", "yolo_mark.vcxproj", "{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Debug|x64.ActiveCfg = Debug|x64
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Debug|x64.Build.0 = Debug|x64
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Debug|x86.ActiveCfg = Debug|Win32
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Debug|x86.Build.0 = Debug|Win32
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Release|x64.ActiveCfg = Release|x64
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Release|x64.Build.0 = Release|x64
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Release|x86.ActiveCfg = Release|Win32
-		{71E25DC4-69FF-4FE2-9DDA-0517AEF2AEAD}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
